console: Refactor device add views
