+++ pkg/webui/console/views/device-add-single/device-add-single.styl
