+++ pkg/webui/console/views/device-add/index.js
@@ -13,28 +13,17 @@
 // limitations under the License.
 
 import React, { Component } from 'react'
-import { Container, Col, Row } from 'react-grid-system'
-import bind from 'autobind-decorator'
 import { connect } from 'react-redux'
-import { push } from 'connected-react-router'
+import { Switch, Route } from 'react-router'
 
 import Breadcrumb from '../../../components/breadcrumbs/breadcrumb'
 import { withBreadcrumb } from '../../../components/breadcrumbs/context'
-import Message from '../../../lib/components/message'
-import IntlHelmet from '../../../lib/components/intl-helmet'
-import DeviceDataForm from '../../components/device-data-form'
 import sharedMessages from '../../../lib/shared-messages'
-import Button from '../../../components/button'
-import withRequest from '../../../lib/components/with-request'
-import { getDeviceTemplateFormats } from '../../store/actions/device-template-formats'
+import NotFoundRoute from '../../../lib/components/not-found-route'
 import { selectSelectedApplicationId } from '../../store/selectors/applications'
-import { selectDeviceTemplateFormats } from '../../store/selectors/device-template-formats'
-import { getDeviceId } from '../../../lib/selectors/id'
-import { selectNsConfig, selectJsConfig, selectAsConfig } from '../../../lib/selectors/env'
 import PropTypes from '../../../lib/prop-types'
-import api from '../../api'
-
-import style from './device-add.styl'
+import DeviceAddSingle from '../device-add-single'
+import DeviceAddBulk from '../device-add-bulk'
 
 @withBreadcrumb('devices.add', function(props) {
   const { appId } = props.match.params
@@ -46,92 +35,21 @@ import style from './device-add.styl'
     />
   )
 })
-@connect(
-  state => ({
-    appId: selectSelectedApplicationId(state),
-    deviceTemplateFormats: selectDeviceTemplateFormats(state),
-    asConfig: selectAsConfig(),
-    nsConfig: selectNsConfig(),
-    jsConfig: selectJsConfig(),
-  }),
-  dispatch => ({
-    redirectToList: (appId, deviceId) =>
-      dispatch(push(`/applications/${appId}/devices/${deviceId}`)),
-    getDeviceTemplateFormats: () => dispatch(getDeviceTemplateFormats()),
-  }),
-)
-@withRequest(({ getDeviceTemplateFormats }) => getDeviceTemplateFormats())
 export default class DeviceAdd extends Component {
   static propTypes = {
     appId: PropTypes.string.isRequired,
-    asConfig: PropTypes.stackComponent.isRequired,
-    deviceTemplateFormats: PropTypes.shape({}).isRequired,
-    jsConfig: PropTypes.stackComponent.isRequired,
-    nsConfig: PropTypes.stackComponent.isRequired,
-    redirectToList: PropTypes.func.isRequired,
-  }
-
-  state = {
-    error: '',
-  }
-
-  @bind
-  async handleSubmit(values) {
-    const { appId } = this.props
-    const { activation_mode, ...device } = values
-
-    return api.device.create(appId, device, {
-      abp: values.activation_mode === 'abp',
-    })
-  }
-
-  @bind
-  handleSubmitSuccess(device) {
-    const { appId, redirectToList } = this.props
-    const deviceId = getDeviceId(device)
-
-    redirectToList(appId, deviceId)
   }
 
   render() {
-    const { error } = this.state
-    const { asConfig, nsConfig, jsConfig, deviceTemplateFormats } = this.props
-    const canBulkCreate = Object.keys(deviceTemplateFormats).length !== 0
-
-    const initialValues = {
-      network_server_address: nsConfig.enabled ? new URL(nsConfig.base_url).hostname : '',
-      application_server_address: asConfig.enabled ? new URL(asConfig.base_url).hostname : '',
-      join_server_address: jsConfig.enabled ? new URL(jsConfig.base_url).hostname : '',
-    }
+    const { appId } = this.props
+    const basePath = `/applications/${appId}/devices/add`
 
     return (
-      <Container>
-        <Row>
-          <Col sm={6}>
-            <IntlHelmet title={sharedMessages.addDevice} />
-            <Message className={style.title} component="h2" content={sharedMessages.addDevice} />
-          </Col>
-          <Col className={style.bulkCreation} sm={6}>
-            <Button.Link
-              message={sharedMessages.bulkCreation}
-              icon="bulk_creation"
-              to="add/bulk"
-              disabled={!canBulkCreate}
-            />
-          </Col>
-        </Row>
-        <Row>
-          <Col className={style.form} lg={8} md={12}>
-            <DeviceDataForm
-              error={error}
-              onSubmit={this.handleSubmit}
-              onSubmitSuccess={this.handleSubmitSuccess}
-              initialValues={initialValues}
-              jsConfig={jsConfig}
-            />
-          </Col>
-        </Row>
-      </Container>
+      <Switch>
+        <Route exact path={basePath} component={DeviceAddSingle} />
+        <Route exact path={`${basePath}/bulk`} component={DeviceAddBulk} />
+        <NotFoundRoute />
+      </Switch>
     )
   }
 }
